Major Restructure of System (2)
This commit is a complete restructure of the entire system, with all nodes,
operators, sockets and menus separated into individual files.

Checks are still on-going, so there may still be some errors in the code that need
investigating.

New features include the ability to animate bones from Live MIDI, but not yet from
control objects made from MIDI files - this is the next job.